+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 7.6.1 (git master)
 * Improve handling of MPP files where MPXJ is unable to read the filter definitions.
+* Improve handling of SureTrak projects without a WBS.
 
 ## 7.6.0 (13/07/2018)
 * Added support for reading ConceptDraw PROJECT CDPX, CPDZ and CPDTZ files.